requirements.txt
# Core MCP (pinned min versions; no latest)
fastmcp>=2.3.0
mcp>=1.6.0

# HTTP & retry
httpx>=0.27.0
tenacity>=8.2.3

# Parsing
tomli>=2.0.1
ruamel.yaml>=0.18.6
packaging>=24.0

# Validation & config
pydantic>=2.7.0
pydantic-settings>=2.3.0

# License analysis
license-expression>=30.3.1

# Env
python-dotenv>=1.0.1
